Move WaitForTasks next to WaitForJobs
