+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/ingest/IngestTasksDriver.java
@@ -15,6 +15,8 @@
  */
 package sleeper.systemtest.dsl.ingest;
 
+import sleeper.systemtest.dsl.util.WaitForTasks;
+
 public interface IngestTasksDriver {
 
     WaitForTasks waitForTasksForCurrentInstance();
