+++ java/system-test/system-test-dsl/src/test/java/sleeper/systemtest/dsl/util/WaitForTasksTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.dsl.ingest;
+package sleeper.systemtest.dsl.util;
 
 import org.junit.jupiter.api.Test;
